movimiento

+++ src/main/DragRosa.java
@@ -5,7 +5,7 @@ import processing.core.PImage;
 
 public class DragRosa extends Drags {
 
-	public DragRosa(PImage imagendrag, String accion, int posx, int posy,float i,float j, PApplet app) {
+	public DragRosa(PImage imagendrag, String accion, int posx, int posy,double i,double j, PApplet app) {
 		super(imagendrag, accion, posx, posy,i,j, app);
 		
 	}
